# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API.",
        "title": "Go Service Demo",
        "contact": {
            "email": "evan@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/ping": {
            "get": {
                "description": "ping for test service alive or not",
                "produces": [
                    "application/json"
                ],
                "summary": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/state/{jobId}": {
            "get": {
                "description": "get job id, name, state, result, error",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "get state of job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "job id",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HttpErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.HttpErrorBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "entity.Job": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "e6e07f18-cae7-4ea9-a4ba-2c8ac364ea5b"
                },
                "name": {
                    "type": "string",
                    "example": "compressImage"
                },
                "result": {},
                "state": {
                    "type": "string",
                    "example": "SUCCESS"
                }
            }
        }
    }
}